Use more resilient way of getting hostname
If app is under a proxy which is trusted, the hostname property
will be set to the X-Forwarded-Host header instead, so it's more
appropriate to be used.

Only using it if available, since on parse.com the hostname
property is undefined.

Also using express's get instead of Connect's headers dict.